fix: load pretrained pipe.pkl instead of retraining on every startup
- Replace train_model() with joblib.load('pipe.pkl') wrapped in @st.cache_resource
- Move training logic to train.py for contributors who need to retrain locally
- Eliminates cold-start CSV loading and XGBoost fitting on Streamlit Cloud
